Consolidate CuPy utility tests
